// requirement/game/frontend/src/types.ts
export type InputMessage = {
  type: 'input';
  direction: 'up' | 'down' | 'stop';
};

export type ResetMessage = {
  type: 'reset';
};

export type StartMessage = {
  type: 'start';
};

export type PauseMessage = {
  type: 'pause';
};

export type StateMessage = {
  type: 'state';
  state: GameState;
  paused: boolean;
};

export type ServerMessage = InputMessage | ResetMessage | StartMessage | PauseMessage | StateMessage;

export type GameState = {
  ball: { x: number; y: number };
  paddles: { p1: number; p2: number };
  score: { p1: number; p2: number };
  ballColor?: string;
  paddleColor?: string;
};